+++ src/Providers/ReviewsIO/Request/GetServiceReviews.php
@@ -1,21 +1,19 @@
 <?php
 
-namespace PodPoint\Reviews\Providers\ReviewsIo\Request\Service;
+namespace PodPoint\Reviews\Providers\ReviewsIO\Request;
 
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Uri;
 use \PodPoint\Reviews\Request\BaseRequest;
 
-class FindReviewRequest extends BaseRequest
+class GetServiceReviews extends BaseRequest
 {
     /**
      * @return array
      */
     public function requiredFields(): array
     {
-        return [
-            'reviewId'
-        ];
+        return ['store'];
     }
 
     /**
@@ -23,11 +21,10 @@ class FindReviewRequest extends BaseRequest
      */
     public function getRequest(): Request
     {
-        $uri = new Uri('https://api.reviews.io/merchant/reviews');
-        $uri = Uri::withQueryValues(
-            $uri,
-            ['store' => $this->getOption('store'), 'review_id' => $this->getOption('reviewId')]
-        );
+        $store = $this->getOption('store');
+
+        $uri = new Uri('https://api.reviews.co.uk/merchant/reviews');
+        $uri = Uri::withQueryValues($uri, $this->options + ['store' => $store]);
 
         return new Request('GET', $uri);
     }
